Update include/matrix/matrix.hpp
Updated first two parameterized constructors to correctly allocate for
the inner vector for the attribute '_data'.

Added two new parameterized constructors one that takes an input of a
vector within an initializer list, and another that takes an initializer
list within an initializer list.

Added an element-wise add method for const variables.

Added an element-wise multiplication method for const variables.

Fixed bug in seq_mult that the inner most loop would inifintely run
until it caused a segmentation fault, it now stops when it is supposed
to.

Added a vector multiply method so that the matrix class can be
multiplied with std::vector.

Added const version of transpose method.

Added operator overloads for '*' operator in the context for multiplying
a matrix with a matrix and a matrix with a vector.